+++ mission/Chapter_10/10/package.json
@@ -13,6 +13,7 @@
     "@tailwindcss/vite": "^4.1.18",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
+    "react-router-dom": "^6.28.0",
     "tailwindcss": "^4.1.18"
   },
   "devDependencies": {
